+++ docs/ChangeLog.md
@@ -157,6 +157,7 @@
 - util (`ble-import`): do not specify arguments to `-C callback` `#D2277` 4f0e94a2
 - main: fix issues with `ble/bin/awk` (reported by devidw) `#D2292` b0a7adcb
 - util (`ble/path#remove-glob`): fix a bug that `*` matches and removes multiple paths `#D2310` xxxxxxxx
+- decode (`bind`): print the filename and line in error messages (motivated by excited-bore) `#D2311` xxxxxxxx
 
 ## Compatibility